refactor encodwe
